requirements.txt
Django==4.2.11
python-dotenv==1.0.0
whitenoise==6.6.0
django-import-export==3.3.7
gunicorn==21.2.0